+++ src/pages/NotFoundPage.tsx
@@ -1,0 +1,9 @@
+import { Link } from "react-router-dom"
+
+export default function NotFoundPage() {
+  return (
+    <div>
+      页面不存在 <Link to="/">回首页</Link>
+    </div>
+  )
+}